Selct the LCD BSB driver
A lot of modification in internal structure.
Anyway, using only the HAL Driver doesn't work. For this reason
I changed my mind and choose the BSB function (that call a lot of things
that I really don't undestood, but work).

I have activated the sdram (with BSP part) modified systemclocl_config and activate LCD.

Next step is to move my framebuffer on SDRAM.

Signed-off-by: Alexjan Carraturo <axjslack@gmail.com>